Fixup some small errors
