+++ entry-generation/godot-entry-generator/src/main/kotlin/godot/entrygenerator/generator/provider/DefaultValueHandlerProvider.kt
@@ -33,7 +33,7 @@ object DefaultValueHandlerProvider {
                 propertyDescriptor,
                 bindingContext
             )
-            propertyDescriptor.type.isCompatibleList() -> Array2RegistrationValuesHandler(
+            propertyDescriptor.type.isCompatibleList() -> ArrayRegistrationValuesHandler(
                 propertyDescriptor,
                 bindingContext
             )
